requirements.txt
asgiref==3.4.1
Brotli==1.0.9
cffi==1.15.0
cssselect2==0.4.1
Django==3.2.9
django-extensions==3.1.5
django-weasyprint==2.1.0
djangorestframework==3.12.4
fonttools==4.28.2
html5lib==1.1
Pillow==8.4.0
pycparser==2.21
pydyf==0.1.2
pyphen==0.11.0
pytz==2021.3
qrcode==7.3.1
six==1.16.0
sqlparse==0.4.2
tinycss2==1.1.1
weasyprint==53.4
webencodings==0.5.1
zopfli==0.1.9
